Consolidate replayable input
